+++ Samples~/Spine/Scripts/SpineAttachmentExample.cs
@@ -1,0 +1,12 @@
+using SaintsField.Spine;
+using Spine.Unity;
+using UnityEngine;
+
+namespace SaintsField.Samples.Spine.Scripts
+{
+    public class SpineAttachmentExample : MonoBehaviour
+    {
+        [SpineAttachment] public string attachmentOriginal;
+        [SpineAttachmentPicker] public string spineAttachment;
+    }
+}